requirements.txt
pytz>=2021.3
python-dateutil>=2.8.2
sympy>=1.9
pytest>=6.0
pytest-cov>=2.0
black>=21.0
flake8>=3.9
sphinx>=4.0
sphinx-rtd-theme>=1.0